Implemented DELETE for messages, also rewrote POST for messages + test coverage
squashed:
Test coverage for DELETE
Minor changes: method & variable renaming, comments
Implemented DELETE for messages, also rewrote POST for messages + test coverage

+++ src/main/java/ru/skillbox/socialnetwork/api/responses/MessageIdResponse.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class MessageIdResponse {
 
   @JsonProperty("message_id")
-  private int messageId;
+  private Long messageId;
 }
